include irqbalance package in automotive
Most of the automotive deployments at this time are expected to be ran
on aarch64-based systems. On these ARM systems with the GIC (Generic
Interrupt Controller), by default all of the interrupts will be
delivered to the first CPU by default. The first CPU can become
overwhelmed with processing interrupts, especially given how busy
some of the automotive boards will be. This was attempted to be
addressed in software in the kernel for aarch64 several times, however
it ran into several issues that resulted in less overall performance,
and increased power consumption as describes at:

https://lore.kernel.org/all/YymRYhjQpRDsD5jM@shell.armlinux.org.uk/

It's recommended to run the irqbalance daemon:

https://lore.kernel.org/all/YymSgJoHEKhD3ypC@shell.armlinux.org.uk/

Signed-off-by: Brian Masney <bmasney@redhat.com>